Update Store and User Service
Add a caching mechanism in the store and update the user service accordingly. If the user service detects that the cache is empty it will fetch twice as many users and put the second half of the in the cache (for future use).

Move the constants file into the shared folder.